Start Makefile and Dockerfile for city web
This isn’t going to be pretty. CityWeb relies
on duktape which has native C extensions.
This means that the common pattern of cross
compiling go binaries for deployment is now busted
and we have to have the docker file do the old
style of fetching go libs via `go get` and compiling
everything.